+++ R/manCULTA-data-results-lta-est.R
@@ -25,6 +25,9 @@
 #'   }
 #'   \item{separation}{
 #'     Level of separation.
+#'     `0` for moderate,
+#'     `-1` for low, and
+#'     `1` for strong.
 #'   }
 #'   \item{n}{
 #'     Sample size.