+++ R/manCULTA-fig-power-lta-est.R
@@ -99,7 +99,11 @@
 FigPowerLTAEst <- function(results_lta_est) {
   Parameters <- power <- Model <- NULL
   results_lta_est$Model <- results_lta_est$model
-  results_lta_est$Separation <- results_lta_est$separation
+  results_lta_est$Separation <- factor(
+    x = results_lta_est$separation,
+    levels = c(-1, 0, 1),
+    labels = c("LO", "MO", "HI")
+  )
   results_lta_est$Parameters <- as.integer(results_lta_est$parnames)
   results_lta_est$n_label <- paste0(
     "N = ",
@@ -139,10 +143,10 @@ FigPowerLTAEst <- function(results_lta_est) {
       n_label ~ Separation
     ) +
     ggplot2::xlab(
-      "Parameters"
+      "Parameter No."
     ) +
     ggplot2::ylab(
-      "Bias"
+      "Statistical Power"
     ) +
     ggplot2::theme_bw() +
     ggplot2::scale_color_brewer(palette = "Set1") +